requirements.txt
aiogram==3.13.1
psycopg2-binary==2.9.9
dateparser==1.2.0
openai==1.58.1
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1